--- src/components/App.jsx
@@ -1,13 +1,0 @@
-import { Header } from "./Header";
-import { HeaderHero } from "./Header-hero";
-
-function App() {
-  return (
-    <div className="body">
-      <Header />
-      <HeaderHero />
-    </div>
-  );
-}
-
-export default App;
