+++ src/components/App/App.css
@@ -1,0 +1,4 @@
+.body {
+    font-family: Inter;
+    font-style: normal;
+}
