+++ src/components/App/App.jsx
@@ -1,0 +1,14 @@
+import "./App.css"
+import { Header } from "../Header/Header";
+import { Offer } from "../Offer/Offer";
+
+function App() {
+  return (
+    <div className="body">
+      <Header />
+      <Offer />
+    </div>
+  );
+}
+
+export default App;
