--- src/components/Header-hero.jsx
@@ -1,9 +1,0 @@
-function HeaderHero() {
-    return (
-        <div className="hero">
-
-        </div>
-    )
-}
-
-export { HeaderHero };
