+++ src/components/Offer/Offer.css
@@ -1,0 +1,4 @@
+.offer {
+    min-height: 648px;
+    background: #465DFF;
+}
