+++ src/components/Offer/Offer.jsx
@@ -1,0 +1,11 @@
+import "./Offer.css";
+
+function Offer() {
+    return (
+        <div className="offer">
+
+        </div>
+    )
+}
+
+export { Offer };
