+++ src/index.css
@@ -1,4 +1,2 @@
 @import url(./vendor/normalize.css);
 @import url(./vendor/fonts.css);
-
-@import url(./blocks/header/header.css);
